[DEVX-1176] Deal -> FxPlugin::Deal
