+++ lib/global_reach_partners.rb
@@ -10,7 +10,6 @@ require 'global_reach_partners/configuration'
 require 'global_reach_partners/currency'
 require 'global_reach_partners/request'
 require 'global_reach_partners/trade_service_request'
-require 'global_reach_partners/deal'
 require 'global_reach_partners/rate'
 require 'global_reach_partners/rate_matrix'
 require 'global_reach_partners/operations'
