+++ lib/global_reach_partners/fx_plugin.rb
@@ -1,1 +1,2 @@
 require 'global_reach_partners/fx_plugin/request'
+require 'global_reach_partners/fx_plugin/deal'
